@@ -1,4 +1,5 @@
-const HtmlWebpackPlugin = require("html-webpack-plugin");
+import HtmlWebpackPlugin from "html-webpack-plugin";
+
 
 const babelLoader = {
   loader: "babel-loader",
@@ -13,44 +14,44 @@ const tsLoader = {
   },
 };
 
+export const entry = {
+  app: "./src/script.tsx",
+};
+export const devServer = {
+  historyApiFallback: true,
+};
+export const resolve = {
+  extensions: [".ts", ".tsx", ".js", ".jsx", ".json"],
+};
+
 module.exports = {
-  entry: {
-    app: "./src/script.tsx",
-  },
-  devServer: {
-    historyApiFallback: true,
-  },
-  resolve: {
-    extensions: [".ts", ".tsx", ".js", ".jsx", ".json"],
-  },
-  module: {
     rules: [
-      {
-        test: /\.scss$/i,
-        use: ["style-loader", "css-loader", "sass-loader"],
-      },
-      {
-        test: /\.css$/i,
-        use: ["style-loader", "css-loader"],
-      },
-      {
-        test: /\.jsx?$/i,
-        use: [babelLoader],
-        exclude: /node_modules/,
-      },
-      {
-        test: /\.tsx?$/i,
-        use: [babelLoader, tsLoader],
-      },
-      {
-        exclude: [/^$/, /\.(js|jsx|ts|tsx)$/i, /\.s?css$/i, /\.html$/i, /\.json$/i],
-        type: "asset/resource",
-      },
-    ],
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: "./src/index.html",
-    }),
+    {
+      test: /\.scss$/i,
+      use: ["style-loader", "css-loader", "sass-loader"],
+    },
+    {
+      test: /\.css$/i,
+      use: ["style-loader", "css-loader"],
+    },
+    {
+      test: /\.jsx?$/i,
+      use: [babelLoader],
+      exclude: /node_modules/,
+    },
+    {
+      test: /\.tsx?$/i,
+      use: [babelLoader, tsLoader],
+      exclude: /node_modules/,
+    },
+    {
+      exclude: [/^$/, /\.(js|jsx|ts|tsx)$/i, /\.s?css$/i, /\.html$/i, /\.json$/i],
+      type: "asset/resource",
+    },
   ],
 };
+export const plugins = [
+  new HtmlWebpackPlugin({
+    template: "./src/index.html",
+  }),
+];
