@@ -1,3 +1,5 @@
-export function div(a,b){ return (
- a/ b)
+export function div(a,b){ 
+    return (
+        a/ b
+    )
 }
